requirements.txt
streamlit
pandas
plotly
langchain
langchain-groq
sentence-transformers
huggingface-hub
python-dotenv
faiss-gpu